pyproject.toml
[project]
name = "parlai"
version = "0.1.0"
requires-python = ">=3.10, < 3.11"
dependencies = [
    "pandas>=1.5,<2.1",
    "numpy>=1.23,<1.26",
    "fastapi[standard]>=0.110",
    "uvicorn>=0.27",
    "pydantic>=2.0,<3.0",
    "scikit-learn>=1.2",
    "autogluon.tabular[all]>=1.1.1",
    "bs4>=0",
    "html5lib>=0"
]

[tool.setuptools.packages.find]
where = ["src"]

